+++ CMAKELISTS.txt
@@ -10,6 +10,15 @@ set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/bin)
 project(managekinetic)
 
+# Define the buildtype.
+if (BUILDTYPE_RELEASE)
+	message("-- [ Config ] Buildtype is release mode.")
+	add_compile_definitions(BUILDTYPE_RELEASE)
+else()
+	message("-- [ Config ] Buildtype is debug mode.")
+	add_compile_definitions(BUILDTYPE_DEBUG)
+endif()
+
 # Define the executable and source files.
 add_executable(managekinetic WIN32
 
@@ -18,16 +27,27 @@ add_executable(managekinetic WIN32
 
 ./src/managekinetic/primitives.h
 
+./src/managekinetic/application.cpp
+./src/managekinetic/application.h
 ./src/managekinetic/async.cpp
 ./src/managekinetic/async.h
 ./src/managekinetic/commandline.cpp
 ./src/managekinetic/commandline.h
 ./src/managekinetic/mappingtable.cpp
 ./src/managekinetic/mappingtable.h
-./src/managekinetic/uicomponent.cpp
-./src/managekinetic/uicomponent.h
+./src/managekinetic/guiwindow.cpp
+./src/managekinetic/guiwindow.h
 ./src/managekinetic/tableviewer.cpp
 ./src/managekinetic/tableviewer.h
+
+./src/managekinetic/GUIwindows/main_window.cpp
+./src/managekinetic/GUIwindows/main_window.h
+./src/managekinetic/GUIwindows/mapping_window.cpp
+./src/managekinetic/GUIwindows/mapping_window.h
+./src/managekinetic/GUIwindows/export_progress_window.cpp
+./src/managekinetic/GUIwindows/export_progress_window.h
+./src/managekinetic/GUIwindows/parse_progress_window.cpp
+./src/managekinetic/GUIwindows/parse_progress_window.h
 
 ./src/platform/fileio_win32.cpp
 ./src/platform/fileio_win32.h
@@ -62,7 +82,7 @@ add_subdirectory(src/vendor/OpenXLSX)
 # Look for OpenGL
 find_package(OpenGL)
 if (OPENGL_FOUND)
-	message("OpenGL Found.")
+	message("-- [ Config ] OpenGL Found.")
 	include_directories(${OPENGL_INCLUDE_DIR})
 endif (OPENGL_FOUND)
 
@@ -76,6 +96,7 @@ endif ()
 # Set Visual Studio's startup project to the base project.
 if (MSVC)
 	set_property(DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR} PROPERTY VS_STARTUP_PROJECT managekinetic)
+	target_compile_options(managekinetic PRIVATE "/MP")
 endif ()
 
 # Allow absolute referencing for project files located in ./src
