+++ src/main.h
@@ -7,9 +7,4 @@
 
 #include <vendor/OpenXLSX/OpenXLSX.hpp>
 
-#define MULTIVALUE_SEPERATOR_UNSIGNED 	253
-#define MULTIVALUE_SEPERATOR_SIGNED 	-3
-#define MULTIVALUE_FIELD_DELIMITER 		'~'
-#define ROW_FIELD_DELIMITER 			';'
-
 #endif
