+++ src/managekinetic/GUIwindows/export_progress_window.h
@@ -1,0 +1,21 @@
+#ifndef MANAGEKINETIC_GUIWIN_EXPORTPROGRESSWINDOW_H
+#define MANAGEKINETIC_GUIWIN_EXPORTPROGRESSWINDOW_H
+#include <managekinetic/guiwindow.h>
+
+class ExportWindow : public GUIWindow
+{
+	public:
+		ExportWindow();
+		~ExportWindow();
+
+	protected:
+		virtual bool onDisplay() 	override;
+		virtual bool onOpen() 		override;
+		virtual bool onClose() 		override;
+		virtual void onExit(bool) 	override;
+
+	private:
+		bool popupOpened = false;
+};
+
+#endif
