+++ src/managekinetic/application.h
@@ -7,7 +7,12 @@
 #include <windows.h>
 #include <managekinetic/primitives.h>
 #include <managekinetic/mappingtable.h>
-#include <managekinetic/uicomponent.h>
+#include <managekinetic/guiwindow.h>
+
+#define MULTIVALUE_SEPERATOR_UNSIGNED 	253
+#define MULTIVALUE_SEPERATOR_SIGNED 	-3
+#define MULTIVALUE_FIELD_DELIMITER 		'~'
+#define ROW_FIELD_DELIMITER 			';'
 
 // Defines all the work that needs to be done.
 typedef std::vector<std::string> UnprocessedTable;
@@ -50,10 +55,6 @@ class document
 		tablemap& 	getTable() { return this->table; };
 		void 		lockTable() { this->tableLock.lock(); }
 		void 		unlockTable() { this->tableLock.unlock(); }
-
-
-	public:
-		i32 currentPreviewPage = 0;
 
 	protected:
 		std::filesystem::path filepath;
