+++ src/managekinetic/mappingtable.h
@@ -7,121 +7,6 @@
 
 #include <managekinetic/primitives.h>
 
-class MappingTable;
-class MappingRecord;
-class MappingField;
-
-class MappingField
-{
-	public:
-		~MappingField() 	{};
-
-		void insertValue(std::string value)
-		{
-			this->values.push_back(value);
-		}
-
-		std::string& operator[](ui64 index)
-		{
-			return this->values[index];
-		}
-
-		ui64 count()
-		{
-			return this->values.size();
-		}
-
-	protected:
-		std::vector<std::string> values;
-		MappingField(MappingTable* parentTable, MappingRecord* parentRecord) 
-			: parentTable(parentTable), parentRecord(parentRecord) {};
-
-	private:
-		MappingTable* parentTable;
-		MappingRecord* parentRecord;
-		friend MappingTable;
-		friend MappingRecord;
-};
-
-class MappingRecord
-{
-	public:
-		~MappingRecord() 	{};
-
-		MappingField& createField()
-		{
-			this->fields.push_back({this->parent, this});
-			return this->fields.back();
-		}
-
-		MappingField& operator[](ui64 index)
-		{
-			return this->fields[index];
-		}
-
-		ui64 count()
-		{
-			return this->fields.size();
-		}
-
-	protected:
-		std::vector<MappingField> fields;
-		MappingRecord(MappingTable* parent) : parent(parent) {};
-
-	private:
-		MappingTable* parent;
-		friend MappingTable;
-};
-
-class MappingTable
-{
-	public:
-		MappingTable() 		{};
-		~MappingTable() 	{};
-
-
-		MappingRecord& createRecord()
-		{
-			this->records.push_back({this});
-			return this->records.back();
-		};
-
-
-		MappingRecord& operator[](ui64 index)
-		{
-			return this->records[index];
-		}
-
-
-		ui64 count()
-		{
-			return this->records.size();
-		}
-
-	protected:
-		std::vector<MappingRecord> records;
-
-};
-
-class MappingDatabase
-{
-	public:
-		MappingDatabase() {};
-		~MappingDatabase()
-		{
-			
-		};
-
-
-
-	protected:
-		std::vector<MappingTable> tables;
-};
-
-// -----------------------------------------------------------------------------
-// Refactor of MappingTable ==> tablemap
-// -----------------------------------------------------------------------------
-
 class tablemap;
 class subtable;
 class record;
